Add files via upload
MySQL script file for use with the Learn SQL Youtube videos.  This is a simple MySQL script that will create the database and tables and populate the tables with sample data.  It contains triggers.  It has one column that is not normalized but it is for demonstration purposes only.  Use as you feel fit.